Refactor Handlers to their own package
